Move markdown_routes to api/markdown
All new API definitions should be migrated to "wqflask/ api"
